wrapping up with job details and restructuring code
